Move images into assets folder, migrate app.tsx code to PXBlueComponents and remove file
