+++ src/pages/contextual-page-templates/Alarms.tsx
@@ -1,3 +1,3 @@
 import React from 'react';
 
-export const Alarms: React.FC = () => <div>Alarms Work!</div>
+export const Alarms: React.FC = () => <div>Alarms Work!</div>;
