+++ src/pages/contextual-page-templates/Timeline.tsx
@@ -1,3 +1,3 @@
 import React from 'react';
 
-export const Timeline: React.FC = () => <div>Timeline Works!</div>
+export const Timeline: React.FC = () => <div>Timeline Works!</div>;
