--- src/pages/kitchen-sink/PXBlueComponents.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const PXBlueComponents: React.FC = () => <div>PXBlueComponents Works!</div>
