--- src/pages/kitchen-sink/index.tsx
@@ -1,2 +1,0 @@
-export * from './PXBlueComponents';
-export * from './material-ui';
