--- src/pages/kitchen-sink/material-ui/DataDisplay.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const DataDisplay: React.FC = () => <div>Data Display Works!</div>
