--- src/pages/kitchen-sink/material-ui/Feedback.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Feedback: React.FC = () => <div>Feedback Works!</div>
