--- src/pages/kitchen-sink/material-ui/Inputs.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Inputs: React.FC = () => <div>Inputs Work!</div>
