--- src/pages/kitchen-sink/material-ui/Navigation.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Navigation: React.FC = () => <div>Navigation Works!</div>
