--- src/pages/kitchen-sink/material-ui/Surfaces.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Surfaces: React.FC = () => <div>Surfaces Work!</div>
