+++ src/pages/material-ui/DataDisplay.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const DataDisplay: React.FC = () => <div>Data Display Works!</div>;
