+++ src/pages/material-ui/Feedback.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Feedback: React.FC = () => <div>Feedback Works!</div>;
