+++ src/pages/material-ui/Inputs.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Inputs: React.FC = () => <div>Inputs Work!</div>;
