+++ src/pages/material-ui/Navigation.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Navigation: React.FC = () => <div>Navigation Works!</div>;
