+++ src/pages/material-ui/Surfaces.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Surfaces: React.FC = () => <div>Surfaces Work!</div>;
